Format.BinaryStore

Description

A Binary Store is a data format that stores a sequence of values encoded using the binary transform. Therefore, it use is restricted to binary lists, i.e. lists whose length is a power of two.

To create a binary store from a binary list use createBinaryStore. This will create a BinaryStore value that you can convert to a lazy ByteString with encode. To revert the process, use decode and then readBinaryStore. The whole process of decoding, from ByteString to the final Decoded value is done lazily. This allows the user to decode only a portion of the data without reading the full ByteString. This is useful when reading big files or when the ByteString is obtained via network connection.

Binary Store Type

data BinaryStore

Binary Store is a format to store data encoded using the Average Binary Transform.

Encoding/Decoding

encode :: BinaryStore -> ByteString

Encode a binary store as a lazy ByteString.

decode :: ByteString -> Either String BinaryStore

Decode a lazy ByteString as a binary store. The header of the binary store is read strictly, while the body evaluation is delayed. It returns a String if the header is malformed. The String contains an error description.

Creating and reading

createBinaryStore

Arguments

:: BinaryStoreValue a

=> Direction
Direction of encoding

-> Word8
Average constant numerator

-> Word8
Average constant denominator

-> Bool
Whether to use zero compression or not

-> Bool
Whether to use bzip compression or not

-> BinList a
Input list

-> Either String BinaryStore

Create a binary store from a binary list, using some configurations. The denominator of the average constant must be greater or equal to its numerator and, of course, different from zero.

createBinaryStoreDefault :: BinaryStoreValue a => BinList a -> BinaryStore

Create a binary store from a binary list, using createBinaryStore with default arguments. In spite of seeming partial (since createBinaryStore returns an Either value), this function is total.

readBinaryStore :: BinaryStoreValue a => BinaryStore -> Decoded a

Read a binary store and build a Decoded value. The Decoded value is a list of partial results of increasing size (1, 2, 4, 8, etc) that ends in either a decoding error or a final result. These partial results are generated lazily from the binary store data.

Class of storable values

class BinaryStoreValue a

This is the class of values that can be stored in a BinaryStore. This is a closed class, so the user is not allowed to add new instances.

Minimal complete definition

putValue, getValue, modeValue, averageBijection

Information

Some functions to get information about a binary store.

data Mode

Binary Store Mode. The Binary Store Mode indicates what kind of data a binary store contains.

Plain: Each value is a Double.

WithHoles: Each value is a Maybe Double.

Constructors

Plain

WithHoles

bsMode :: BinaryStore -> Mode

Binary Store mode.

bsNumerator :: BinaryStore -> Word8

Numerator of the Average Constant.

bsDenominator :: BinaryStore -> Word8

Denominator of the Average Constant.

averageConstant :: BinaryStore -> Double

The constant used by the Average Binary Transform.

bsDirection :: BinaryStore -> Direction

Direction of encoding.

bsCompression :: BinaryStore -> Bool

Whether zero compression is used or not.

bsBZip :: BinaryStore -> Bool

Wheter bzip compression is used or not.

bsLength :: BinaryStore -> Word8

Length index of the data.

bsData :: BinaryStore -> ByteString

Data (which might be compressed).